+++ src/com/kstenschke/referencer/Preferences.java
@@ -32,6 +32,8 @@ public class Preferences {
     @NonNls private static final String PROPERTY_REFERENCER_INDEX_PHP = "PluginReferencer.SelectedIndexPHP";
     @NonNls private static final String PROPERTY_REFERENCER_INDEX_JS = "PluginReferencer.SelectedIndexJS";
     @NonNls private static final String PROPERTY_GOTO_LIST_BOOKMARKS = "PluginReferencer.GotoListBookmarks";
+    @NonNls private static final String PROPERTY_GOTO_LIST_PHP_METHODS = "PluginReferencer.GotoListPhpMethods";
+    @NonNls private static final String PROPERTY_GOTO_LIST_JS_METHODS = "PluginReferencer.GotoListJsMethods";
 
     public static void saveGoToPatterns(String patterns) {
         PropertiesComponent.getInstance().setValue(PROPERTY_PATTERNS_GOTO, patterns);
@@ -43,6 +45,14 @@ public class Preferences {
 
     public static void saveShowBookmarksInGoToList(boolean doShow) {
         PropertiesComponent.getInstance().setValue(PROPERTY_GOTO_LIST_BOOKMARKS, doShow ? "1" : "0");
+    }
+
+    public static void saveShowPhpMethodsInGoToList(boolean doShow) {
+        PropertiesComponent.getInstance().setValue(PROPERTY_GOTO_LIST_PHP_METHODS, doShow ? "1" : "0");
+    }
+
+    public static void saveShowJsMethodsInGoToList(boolean doShow) {
+        PropertiesComponent.getInstance().setValue(PROPERTY_GOTO_LIST_JS_METHODS, doShow ? "1" : "0");
     }
 
     public static String getGoToPatterns() {
@@ -58,6 +68,18 @@ public class Preferences {
     }
 
     public static boolean getShowBookmarksInGoToList() {
+        String doShow = PropertiesComponent.getInstance().getValue(PROPERTY_GOTO_LIST_BOOKMARKS);
+
+        return doShow != null && doShow.equals("1");
+    }
+
+    public static boolean getShowPhpMethodsInGoToList() {
+        String doShow = PropertiesComponent.getInstance().getValue(PROPERTY_GOTO_LIST_BOOKMARKS);
+
+        return doShow != null && doShow.equals("1");
+    }
+
+    public static boolean getShowJsMethodsInGoToList() {
         String doShow = PropertiesComponent.getInstance().getValue(PROPERTY_GOTO_LIST_BOOKMARKS);
 
         return doShow != null && doShow.equals("1");