+++ src/com/kstenschke/referencer/ReferencerConfigurable.java
@@ -53,6 +53,8 @@ public class ReferencerConfigurable implements Configurable {
     @Override public void apply() {
         if (settingsPanel != null) {
             Preferences.saveShowBookmarksInGoToList(settingsPanel.getShowBookmarksInGotoDestinations());
+            Preferences.saveShowJsMethodsInGoToList(settingsPanel.getShowJsMethodsInGotoDestinations());
+            Preferences.saveShowPhpMethodsInGoToList(settingsPanel.getShowPhpMethodsInGotoDestinations());
             Preferences.saveGoToPatterns(settingsPanel.getGoToPatterns());
             Preferences.saveReplacePatterns(settingsPanel.getReplacePatterns());
         }